+++ packages/okta-react/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### Bug Fixes
 
-- Adjusts `<SecureRoute>` so that it enforces authenticationRequirement for components passed via "render" or "children" in addition to "component"
+- [#872](https://github.com/okta/okta-oidc-js/pull/872) Adjusts `<SecureRoute>` so that it enforces authentication requirement for components passed via "render" or "children" in addition to "component"
+  - NOTE: `<SecureRoute>`, like react-router `<Route>`, only wants ONE of the three ways of passing wrapped components per route
 
 # 3.0.4